ClientSideAplication/requirements.txt
# Extensiones necesarias para usar el programa

psutil
pyinstaller
Pillow
pynput